Fix time zone handling for VictorOps updates
The combination of the native JS Date object with
moment-timezone, resulted in a bug with the time zone
handling implemented in PR #422. The result was that the
time stamp (which was already in BST) was converted from
UTC to BST, resulting in an erroneous 1 hour offset in
the VictorOps schedules. Simplified and fixed this by only
using moment and moment-timezone.

Change-type: patch
Signed-off-by: AlidaOdendaal <alida@balena.io>