Create separate nocover agents, one with 2s and one with 3s
In the original logs/example/nocover.json file, the availability was
marked as 3 for the whole week. Therefore, this file could not
be used for the case where we want the teamwork-scheduler to run
NOCOVER="true" and USETHREES="false". A nocover2s.json file was
created for this purpose, and the original file has been renamed
to nocover3s.json.

Change-type: minor
Signed-off-by: AlidaOdendaal <alida@balena.io>