Extend scheduler to allow multiple shifts per agent per day
This change reconfigures the model variables by introducing another
dimension k, which allows an agent to be scheduled in at most
maxShiftsPerAgentPerDay (as specified in the options) shifts
on any given weekday. The penalty function is also extended so as
to bias the scheduler towards preferring not to assign multiple
shifts per day if it's not needed to avoid other costs.

Fibery improvement: https://balena.fibery.io/Work/Improvement/Improve-devOps-scheduler-to-handle-split-shifts-2763

Change-type: major
Signed-off-by: AlidaOdendaal <alida@balena.io>